完成了注册店铺的功能,可同时提交表单数据和上传图片;此次开发中遇到四个问题,其中三个问题在网络上搜索不到有效的解决犯法,耗费了较多时间才解决,

+++ src/main/java/com/kingguanzhang/pojo/Shop.java
@@ -11,7 +11,7 @@ public class Shop {
 
     private String shopAddr;
 
-    private Integer phone;
+    private Long phone;
 
     private String shopImg;
 
@@ -63,11 +63,11 @@ public class Shop {
         this.shopAddr = shopAddr == null ? null : shopAddr.trim();
     }
 
-    public Integer getPhone() {
+    public Long getPhone() {
         return phone;
     }
 
-    public void setPhone(Integer phone) {
+    public void setPhone(Long phone) {
         this.phone = phone;
     }
